+++ src/disclosures/js/views/financial-view.js
@@ -20,6 +20,7 @@ var financialView = {
   $programLength: $( '#estimated-years-attending' ),
   $aboutThisTool: $( '.instructions_about a' ),
   $addPrivateButton: $( '.private-loans_add-btn' ),
+  $totalDirectCostSection: $( '.verify_direct-cost' ),
   $gradPlusSection: $( '[data-section="gradPlus"]' ),
   $perkinsSection: $( '[data-section="perkins"]' ),
   $privateContainer: $( '.private-loans' ),
@@ -42,6 +43,7 @@ var financialView = {
     this.removePrivateListener();
     this.resetPrivateLoanView();
     this.continueStep2Listener();
+    this.termToggleListener();
   },
 
   /**
@@ -101,9 +103,9 @@ var financialView = {
     this.updateLeftovers( values, $leftovers );
     this.updatePrivateLoans( values, $privateLoans );
     this.updateRemainingCostContent();
-    metricView.updateDebtBurden( values );
-    metricView.updateMonthlyPayment();
+    metricView.updateDebtBurden();
     this.updateCalculationErrors( values );
+    this.termToggleVisible( values );
   },
 
   /**
@@ -219,6 +221,15 @@ var financialView = {
     // Update availability of gradPLUS loans
     this.gradPlusVisible( values.level.indexOf( 'Graduate' ) !== -1 );
     this.perkinsVisible( values.offersPerkins );
+  },
+
+  /**
+   * Updates view based on URL values.
+   * Currently updates the visibility of the anticipated total direct cost.
+   * @param {object} values - An object with URL values
+   */
+  updateViewWithURL: function( values ) {
+    this.totalDirectCostVisible( values.totalCost !== undefined );
   },
 
   /**
@@ -443,6 +454,19 @@ var financialView = {
     } );
   },
 
+  /**
+   * Sets visibility of anticipated total direct cost section
+   * Dependent on `totl` value being passed in URL
+   * @param {boolean} visibility - Whether or not `values.totalCost` is not null
+   */
+  totalDirectCostVisible: function( visibility ) {
+    if ( visibility === false ) {
+      this.$totalDirectCostSection.hide();
+    } else {
+      this.$totalDirectCostSection.show();
+    }
+  },
+
   gradPlusVisible: function( visibility ) {
     if ( visibility === false ) {
       this.$gradPlusSection.hide();
@@ -457,6 +481,26 @@ var financialView = {
       this.$perkinsSection.hide();
     } else {
       this.$perkinsSection.show();
+    }
+  },
+
+  termToggleVisible: function( values ) {
+    var fedTotal;
+
+    fedTotal = values.perkinsDebt + values.directSubsidizedDebt;
+    fedTotal += values.directUnsubsidizedDebt + values.gradPlusDebt;
+
+    // If federal loan debt at graduation exceeds $30,000, then
+    // the 25-year repayment term is an option
+    if ( fedTotal > 30000 ) {
+      $( '[data-term-toggle]' ).show();
+      $( '.repaymentContent' ).hide();
+    } else {
+      $( '[data-term-toggle]' ).hide();
+      if ( values.repaymentTerm !== 10 ) {
+        publish.financialData( 'repaymentTerm', 10 );
+        financialView.updateView( getFinancial.values() );
+      }
     }
   },
 
@@ -487,6 +531,21 @@ var financialView = {
       .on( 'sticky_kit:unbottom', function( e ) {
         $( e.target ).removeClass( 'is_bottomed' );
       } );
+  },
+
+  /**
+   * Listener for clicks on the repayment toggles
+   */
+  termToggleListener: function() {
+    $( '[data-repayment-section] input' ).click( function() {
+      var $ele = $( this ),
+          $toggles = $( '[data-repayment-section] input' ),
+          term = $ele.val();
+      publish.financialData( 'repaymentTerm', term );
+      $toggles.prop( 'checked', false );
+      $toggles.filter( '[value="' + term + '"]' ).prop( 'checked', true );
+      financialView.updateView( getFinancial.values() );
+    } );
   }
 };