+++ CHANGELOG.md
@@ -4,6 +4,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ## Unreleased
 - Error checking added to program view
 - Error checking added to verification view
+- removed 6 unneeded fields from school json payload
 
 ## 2.1.2
 - Parent PLUS loans separated from other family contributions