+++ CHANGELOG.md
@@ -3,6 +3,8 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
 - Added load tests
+- Error checking added to program view
+- Error checking added to verification view
 - removed 6 unneeded fields from school json payload
 
 ## 2.1.2