+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Error checking added to program view
 - Error checking added to verification view
 - removed 6 unneeded fields from school json payload
+- Added notifications script and 2 fields to Notification model for errors
 
 ## 2.1.2
 - Parent PLUS loans separated from other family contributions