+++ CHANGELOG.md
@@ -7,6 +7,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Error checking added to verification view
 - removed 6 unneeded fields from school json payload
 - Added notifications script and 2 fields to Notification model for errors
+- Made program code required/program name not required in load programs script
 
 ## 2.1.2
 - Parent PLUS loans separated from other family contributions