+++ CHANGELOG.md
@@ -2,7 +2,11 @@ All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
+- Added load tests
+- Error checking added to program view
+- Error checking added to verification view
 - removed 6 unneeded fields from school json payload
+- Made program code required/program name not required in load programs script
 
 ## 2.1.2
 - Parent PLUS loans separated from other family contributions