+++ CHANGELOG.md
@@ -11,6 +11,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Made school name bold
 - Use total direct cost value from offer URL
 - Update toggle functionality, interactions, and dynamic content
+- Addressed security vulnerability report in NPM dependencies
 
 ## 2.1.2
 - Parent PLUS loans separated from other family contributions