+++ CHANGELOG.md
@@ -2,6 +2,8 @@ All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
+
+## 2.1.3
 - Added load tests
 - Error checking added to program view
 - Error checking added to verification view